Refactor of the code.

+++ src/main/java/com/shop/repositories/ShopDatabaseRepository.java
@@ -2,5 +2,5 @@ package com.shop.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShopDatabaseRepository extends CrudRepository<ShopEntity, Integer> {
+public interface ShopDatabaseRepository extends CrudRepository<ShopEntity, Long> {
 }
